Step 4
Extract the data from the original mockup and store them in an array of
object literals. What are the objects properties? `{imageUrl:, title:,
comment:, isLiked:}` Add an array of these object literals to the top of
your script. Make a commit!! [Extra credit: To view the mockup, try
going back two commits to the original HTML mockup commit. Use the
github app to get the commit hash of the initial commit, then try `git
checkout <commitHash>` in the terminal (make sure you're in the project
directory). When you're done use `git checkout master` to go back to the
`HEAD` of your repo.]